r move intellij reporters from windows package to its own.
