+++ approvaltests/src/main/java/org/approvaltests/reporters/windows/WindowsDiffReporter.java
@@ -1,6 +1,8 @@
 package org.approvaltests.reporters.windows;
 
 import org.approvaltests.reporters.FirstWorkingReporter;
+import org.approvaltests.reporters.intellij.IntelliJCommunityReporter;
+import org.approvaltests.reporters.intellij.IntelliJUltimateReporter;
 
 public class WindowsDiffReporter extends FirstWorkingReporter
 {
